feat(list-filters): support __none__ sentinel in area filter for work items and household items
Lifts shared resolveAreaFilter helper to areaService; both work-item and
household-item services accept the __none__ sentinel to include null-area
records. Household items upgraded to CSV multi-select + descendant expansion.

Fixes #1277
