+++ server/src/services/workItemService.ts
@@ -1,5 +1,5 @@
 import { randomUUID } from 'node:crypto';
-import { eq, sql, and, or, desc, asc, inArray } from 'drizzle-orm';
+import { eq, sql, and, or, desc, asc, inArray, isNull } from 'drizzle-orm';
 import type { BetterSQLite3Database } from 'drizzle-orm/better-sqlite3';
 import type * as schemaTypes from '../db/schema.js';
 import type { areas } from '../db/schema.js';
@@ -16,7 +16,7 @@ import {
 import { listWorkItemBudgets } from './workItemBudgetService.js';
 import { autoReschedule } from './schedulingEngine.js';
 import { deleteLinksForEntity } from './documentLinkService.js';
-import { getDescendantIds, loadAreaMap, resolveAreaAncestors } from './areaService.js';
+import { getDescendantIds, loadAreaMap, resolveAreaAncestors, resolveAreaFilter } from './areaService.js';
 import type { AreaMapEntry } from './areaService.js';
 import {
   onWorkItemStatusChanged,
@@ -41,28 +41,6 @@ import type {
 import { NotFoundError, ValidationError } from '../errors/AppError.js';
 
 type DbType = BetterSQLite3Database<typeof schemaTypes>;
-
-/**
- * Parse the areaId filter (single ID, CSV string, or array) into an expanded,
- * deduplicated array of area IDs including each supplied ID's descendants.
- * Empty segments and unknown IDs are silently ignored (inArray on unknown IDs returns no rows).
- */
-function resolveAreaIds(db: DbType, areaId: string | string[]): string[] {
-  const raw = Array.isArray(areaId)
-    ? areaId
-    : areaId
-        .split(',')
-        .map((s) => s.trim())
-        .filter((s) => s.length > 0);
-
-  const expanded = new Set<string>();
-  for (const id of raw) {
-    for (const descendant of getDescendantIds(db, id)) {
-      expanded.add(descendant);
-    }
-  }
-  return [...expanded];
-}
 
 /**
  * Count budget lines for a work item.
@@ -656,8 +634,12 @@ export function listWorkItems(
   }
 
   if (query.areaId) {
-    const areaIds = resolveAreaIds(db, query.areaId);
-    if (areaIds.length > 0) {
+    const { areaIds, includeNull } = resolveAreaFilter(db, query.areaId);
+    if (includeNull && areaIds.length > 0) {
+      baseConditions.push(or(isNull(workItems.areaId), inArray(workItems.areaId, areaIds))!);
+    } else if (includeNull) {
+      baseConditions.push(isNull(workItems.areaId));
+    } else if (areaIds.length > 0) {
       baseConditions.push(inArray(workItems.areaId, areaIds));
     }
   }
