+++ shared/src/types/householdItem.ts
@@ -256,6 +256,13 @@ export interface HouseholdItemListQuery {
   q?: string;
   category?: HouseholdItemCategory;
   status?: HouseholdItemStatus;
+  /**
+   * Filter by area UUID or comma-separated list of area UUIDs.
+   * For each supplied ID the server expands descendants server-side;
+   * results are the union of all matching subtrees.
+   * Use the sentinel value `__none__` to include items with no area assignment.
+   * `__none__` can be combined with UUIDs (e.g., `__none__,kitchen-uuid`) for union.
+   */
   areaId?: string;
   vendorId?: string;
   plannedCostMin?: number;
