+++ shared/src/types/workItem.ts
@@ -174,6 +174,13 @@ export interface WorkItemListQuery {
   status?: WorkItemStatus;
   assignedUserId?: string;
   assignedVendorId?: string;
+  /**
+   * Filter by area UUID or comma-separated list of area UUIDs.
+   * For each supplied ID the server expands descendants server-side;
+   * results are the union of all matching subtrees.
+   * Use the sentinel value `__none__` to include items with no area assignment.
+   * `__none__` can be combined with UUIDs (e.g., `__none__,kitchen-uuid`) for union.
+   */
   areaId?: string;
   q?: string;
   budgetLinesMin?: number;
